# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Control de Capital Humano API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List all audited zones
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
    post:
      operationId: createZone
      tags: [zones]
      summary: Create a zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
  /zones/{id}:
    get:
      operationId: getZone
      tags: [zones]
      summary: Get zone by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
    patch:
      operationId: updateZone
      tags: [zones]
      summary: Update zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneUpdate"
      responses:
        "200":
          description: Updated zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
    delete:
      operationId: deleteZone
      tags: [zones]
      summary: Delete zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PANELES
...
  /visual-zones:
    get:
      operationId: listVisualZones
      tags: [visualZones]
      summary: List all visual zones
      responses:
        "200":
          description: List of visual zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisualZone"
    post:
      operationId: createVisualZone
      tags: [visualZones]
      summary: Create a visual zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisualZoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisualZone"
  /visual-zones/{id}:
    get:
      operationId: getVisualZone
      tags: [visualZones]
      summary: Get visual zone by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: VisualZone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisualZone"
    patch:
      operationId: updateVisualZone
      tags: [visualZones]
      summary: Update visual zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisualZoneUpdate"
      responses:
        "200":
          description: Updated visual zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisualZone"
    delete:
      operationId: deleteVisualZone
      tags: [visualZones]
      summary: Delete visual zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # AUDIT CAPTURES
...
components:
  schemas:
...
    Zone:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        udn_id:
          type: ["integer", "null"]
        created_at:
          type: string
          format: date-time
    ZoneInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        udn_id:
          type: integer
    ZoneUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        udn_id:
          type: integer

    # PANEL
...
    VisualZone:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
    VisualZoneInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    VisualZoneUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    # ALPHANUMERIC
...